Organize repositories/ into helpers subfolder; drop dead enum re-export
- Move non-repository modules (types, serializers, ids, user-auth,
  reminder-schedule, duplicate-key-error) into api/repositories/helpers/
  so the directory top level holds only the *.repository.ts files and the
  useRepositories.ts aggregator. Update all internal and external imports.
- Remove the unused StartPages / PassageNoteTagSortOrders re-export from
  UserSettings.ts; the canonical source is validation/constants.ts. The
  schema keeps importing them for its own validators.

No behavior changes.
