+++ api/repositories/helpers/user-auth.ts
@@ -1,5 +1,5 @@
 import jwt from 'jsonwebtoken';
-import config from '../config';
+import config from '../../config';
 import { UserRecord } from './types';
 
 /**
