+++ api/router/routes/passage-notes.ts
@@ -2,9 +2,9 @@ import express from 'express';
 import authCurrentUser from '../helpers/authCurrentUser';
 import { Bible } from '@mybiblelog/shared';
 import useRepositories from '../../repositories/useRepositories';
-import { isValidObjectId } from '../../repositories/ids';
-import { toPassageNoteJSON } from '../../repositories/serializers';
-import { type PassageNoteSearchQuery } from '../../repositories/types';
+import { isValidObjectId } from '../../repositories/helpers/ids';
+import { toPassageNoteJSON } from '../../repositories/helpers/serializers';
+import { type PassageNoteSearchQuery } from '../../repositories/helpers/types';
 import { ApiErrorDetailCode } from '../errors/error-codes';
 import { type ApiResponse } from '../response';
 import { ValidationError } from '../errors/validation-errors';
